Move appAccessRkey to internal/syntax
Exports it as AppAccessRkey alongside AppAccessCollection, which
already lives in internal/syntax — the rkey derivation and the
collection it addresses are the same concern. Updates the two call
sites (GetSpaceCredential's attestation check, the GrantAppAccess
test helper) to call it via the habitat_syntax import already present
in both files.
