+++ internal/syntax/app_access.go
@@ -1,4 +1,4 @@
-package pearserver
+package syntax
 
 import (
 	"encoding/base64"
@@ -7,10 +7,10 @@ import (
 	"github.com/bluesky-social/indigo/atproto/syntax"
 )
 
-// appAccessRkey deterministically derives the network.habitat.space.appAccess
+// AppAccessRkey deterministically derives the network.habitat.space.appAccess
 // record key for a client_id: its base64url (no padding) encoding, so the
 // grant for a given client is directly addressable without a lookup index.
-func appAccessRkey(clientID string) (syntax.RecordKey, error) {
+func AppAccessRkey(clientID string) (syntax.RecordKey, error) {
 	encoded := base64.RawURLEncoding.EncodeToString([]byte(clientID))
 	rkey, err := syntax.ParseRecordKey(encoded)
 	if err != nil {
